requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
pandas
pillow
pytesseract
python-docx
chromadb
faiss-cpu
langchain_google_genai
langchain-community
openpyxl
